logic for user's stats + fixing bug in lobby + chaning for-in to for-of

+++ install/database_schema.sql
@@ -11,8 +11,8 @@ CREATE TABLE users (
     username          NVARCHAR(19)  NOT NULL  UNIQUE, /* MySQL is case insensitive by default, which is what we want */
     password          CHAR(64)      NOT NULL, /* A SHA-256 hash string is 64 characters long */
     num_played        INT           NOT NULL  DEFAULT 0,
-    average_score     INT           NOT NULL  DEFAULT 0,
-    loss_percent      INT           NOT NULL  DEFAULT 0,
+    average_score     FLOAT         NOT NULL  DEFAULT 0,
+    loss_percent      FLOAT         NOT NULL  DEFAULT 0,
     datetime_created  TIMESTAMP     NOT NULL, /* Defaults to the current time */
     last_login        TIMESTAMP     NOT NULL /* Defaults to the current time */
 );
